requirements.txt
tweepy==4.10.0
google-cloud-pubsub==2.13.6
google-cloud-secret-manager